moved texture related functions to their own file
