+++ inc/CBDraycaster.h
@@ -33,5 +33,6 @@ bool	get_mlx(t_raycaster *raycaster);
 bool	get_image(t_raycaster *raycaster);
 bool	load_textures(t_raycaster *raycaster, const t_param param);
 void	cub3d_math_hook(void *parameter);
+void	ray_texture_calc(t_ray ray, const int x, const t_raycaster raycaster);
 
 #endif // !CBDRAYCASTER_H
